requirements.txt
crewai
crewai_tools
load_dotenv
streamlit
pysqlite3-binary
boto3